created VisibilityButton component
